+++ CHANGELOG.md
@@ -9,5 +9,7 @@
 - Escape now closes chat
 - Enter now closes chat if input field has no text
 - Chat message now displays is player dead
+- Chat now hides per
+- Chat messages are now fade separately from each other
 
 ### Other